+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Rebalanced Blasting. The bonus mining speed is now based off tool's mining speed
 * Increased meat drop from animals
 * Golden shield now blocks more damage (1 -> 3)
 * Copper tools now last 3% more per block below sea level (+~200% at y=0 and +~350% at y=-56)